POPUG-014 Refactoring: TaskInfo -> TaskReplicationInfo.
