+++ employee-billing/src/main/java/org/uber/popug/employee/billing/domain/task/replication/TaskReplicationInfo.java
@@ -1,11 +1,11 @@
-package org.uber.popug.employee.billing.domain.task;
+package org.uber.popug.employee.billing.domain.task.replication;
 
 import jakarta.annotation.Nonnull;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-public record TaskInfo(
+public record TaskReplicationInfo(
         @Nonnull UUID id,
         @Nonnull UUID assigneeId,
         @Nonnull String description,
