+++ employee-billing/src/main/java/org/uber/popug/employee/billing/mapping/TasksCUDKafkaEventMapper.java
@@ -2,7 +2,7 @@ package org.uber.popug.employee.billing.mapping;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import org.uber.popug.employee.billing.domain.task.TaskInfo;
+import org.uber.popug.employee.billing.domain.task.replication.TaskReplicationInfo;
 import org.uber.popug.employee.billing.kafka.event.cud.TaskCreatedReplicationEvent;
 
 @Mapper
@@ -10,6 +10,6 @@ public interface TasksCUDKafkaEventMapper {
 
     @Mapping(source = "taskId", target = "id")
     @Mapping(source = "taskDescription", target = "description")
-    TaskInfo toBusiness(TaskCreatedReplicationEvent taskCreatedReplicationEvent);
+    TaskReplicationInfo toBusiness(TaskCreatedReplicationEvent taskCreatedReplicationEvent);
 
 }
