+++ employee-billing/src/main/java/org/uber/popug/employee/billing/service/TaskBillingAssignmentService.java
@@ -1,10 +1,10 @@
 package org.uber.popug.employee.billing.service;
 
 import org.uber.popug.employee.billing.domain.aggregates.TaskWithAssignee;
-import org.uber.popug.employee.billing.domain.task.TaskInfo;
+import org.uber.popug.employee.billing.domain.task.replication.TaskReplicationInfo;
 
 public interface TaskBillingAssignmentService {
 
-    TaskWithAssignee assembleTaskWithAssignee(TaskInfo taskInfo);
+    TaskWithAssignee assembleTaskWithAssignee(TaskReplicationInfo taskReplicationInfo);
 
 }
